optimize target-pool scanning
Use a one time call to gcloud to pull all target-pools and their node
targets as a giant json blob, then some shell+jq magic to end up with
actual objects we can parse without having to describe the pools in a
loop.

Additionally verify that the targets in the target pool belong to the
cluster that we are told to manage. This should avoid accidentally
removing a pool that doesn't belong to this cluster